remove `addition.jl` file (no longer needed)
