+++ src/HMatrices.jl
@@ -61,7 +61,6 @@ include("rkmatrix.jl")
 include("compressor.jl")
 include("hmatrix.jl")
 include("dhmatrix.jl")
-include("addition.jl")
 include("multiplication.jl")
 include("triangular.jl")
 include("lu.jl")
