+++ test/runtests.jl
@@ -24,10 +24,6 @@ end
     include("hmatrix_test.jl")
 end
 
-# @safetestset "Addition" begin
-#     include("addition_test.jl")
-# end
-
 @safetestset "Multiplication" begin
     include("multiplication_test.jl")
 end
